Init commit

+++ Lab3/time_log.c++
@@ -1,16 +1,19 @@
-#include <iostream>      // Для вывода ошибок (std::cerr)
-#include <fstream>       // Для работы с файлом лога (std::ofstream)
-#include <string>        // Для работы со std::string
-#include <chrono>        // Для измерения времени (modern C++)
-#include <iomanip>       // Для форматированного вывода числа (например, leading zeros)
-#include <ctime>         // Для разбора времени на компоненты
+#include <iostream>      // std::cerr
+#include <fstream>       // std::ofstream
+#include <string>        // std::string
+#include <sstream>       // std::ostringstream
+#include <chrono>
+#include <iomanip>       // std::setfill, std::setw
+#include <ctime>         // std::tm
 
 #ifdef _WIN32
-#include <windows.h>     // Для получения времени и PID в Windows
+#include <windows.h>     // SYSTEMTIME, GetSystemTime, GetCurrentProcessId
 #else
-#include <unistd.h>      // Для getpid() на POSIX
-#include <sys/time.h>    // Для gettimeofday() на POSIX
+#include <unistd.h>      // getpid
+#include <sys/time.h>    // gettimeofday
 #endif
+
+#include "time_log.hpp"  // объявления time_to_str и do_log
 
 // Кроссплатформенная функция формирования строки времени с точностью до миллисекунд
 std::string time_to_str() {
@@ -22,13 +25,13 @@ std::string time_to_str() {
 
     // Форматируем дату и время с ведущими нулями
     oss << std::setfill('0')
-        << std::setw(4) << st.wYear << '-'              // Год
-        << std::setw(2) << st.wMonth << '-'             // Месяц
-        << std::setw(2) << st.wDay << ' '               // День
-        << std::setw(2) << st.wHour << ':'              // Часы
-        << std::setw(2) << st.wMinute << ':'            // Минуты
-        << std::setw(2) << st.wSecond << '.'            // Секунды
-        << std::setw(3) << st.wMilliseconds;            // Миллисекунды
+        << std::setw(4) << st.wYear << '-'
+        << std::setw(2) << st.wMonth << '-'
+        << std::setw(2) << st.wDay << ' '
+        << std::setw(2) << st.wHour << ':'
+        << std::setw(2) << st.wMinute << ':'
+        << std::setw(2) << st.wSecond << '.'
+        << std::setw(3) << st.wMilliseconds;
 #else
     struct timeval tv;
     gettimeofday(&tv, nullptr); // Текущее время, включая микросекунды
@@ -38,13 +41,13 @@ std::string time_to_str() {
 
     // Форматирование аналогично Windows
     oss << std::setfill('0')
-        << std::setw(4) << t.tm_year + 1900 << '-'      // Год
-        << std::setw(2) << t.tm_mon + 1 << '-'          // Месяц
-        << std::setw(2) << t.tm_mday << ' '             // День
-        << std::setw(2) << t.tm_hour << ':'             // Часы
-        << std::setw(2) << t.tm_min << ':'              // Минуты
-        << std::setw(2) << t.tm_sec << '.'              // Секунды
-        << std::setw(3) << (tv.tv_usec / 1000);         // Миллисекунды
+        << std::setw(4) << t.tm_year + 1900 << '-'
+        << std::setw(2) << t.tm_mon + 1 << '-'
+        << std::setw(2) << t.tm_mday << ' '
+        << std::setw(2) << t.tm_hour << ':'
+        << std::setw(2) << t.tm_min << ':'
+        << std::setw(2) << t.tm_sec << '.'
+        << std::setw(3) << (tv.tv_usec / 1000);
 #endif
     return oss.str();
 }
@@ -53,17 +56,16 @@ std::string time_to_str() {
 void do_log(const std::string& data) {
     std::ofstream file("log.txt", std::ios::app); // Открываем лог-файл для дозаписи
     if (!file.is_open()) {
-        std::cerr << "Failed to open log file" << std::endl; // Сообщение об ошибке, если не открылся
+        std::cerr << "Failed to open log file" << std::endl;
         return;
     }
 
 #ifdef _WIN32
-    int pid = GetCurrentProcessId(); // Получаем идентификатор процесса на Windows
+    int pid = static_cast<int>(GetCurrentProcessId());
 #else
-    int pid = getpid();              // Получаем идентификатор процесса на POSIX
+    int pid = static_cast<int>(getpid());
 #endif
 
     // Пишем в лог: [PID] время: текст лога
     file << "[" << pid << "] " << time_to_str() << ": " << data << std::endl;
-    // std::ofstream сам закрывает файл и сбрасывает поток при завершении функции
 }
